Rename UniversalCell to ExecutableCell and clean up unused files
- Rename UniversalCell → ExecutableCell (better semantic meaning)
- Update all imports and interface names to match
- Delete unused cell components: CodeCell.tsx, AiCell.tsx, SqlCell.tsx
- Delete unused MarkdownToolbar.tsx (empty placeholder)

Architecture now clearly separates:
- ExecutableCell: Code, AI, SQL cells that can execute
- MarkdownCell: Static content display with edit/preview

Eliminated ~1000 lines of duplicated code while maintaining all functionality.
