--- src/components/notebook/cell/toolbars/MarkdownToolbar.tsx
@@ -1,16 +1,0 @@
-import React from "react";
-// import { Badge } from "@/components/ui/badge";
-
-export const MarkdownToolbar: React.FC = () => {
-  return (
-    <div className="flex items-center gap-2">
-      {/* <Badge
-        variant="outline"
-        className="h-5 border-gray-200 bg-gray-50/50 text-xs text-gray-600"
-        title="Markdown cell"
-      >
-        markdown
-      </Badge> */}
-    </div>
-  );
-};
